devops/assign3/backend/requirements.txt
flask
pymongo[srv]==4.7
dnspython
python-dotenv